change tmmst to ulong

+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoRaWan.NetworkServer/LoRaDeviceTelemetry.cs
@@ -16,7 +16,7 @@ namespace LoRaWan.NetworkServer
         public string Time { get; set; }
 
         [JsonProperty("tmms")]
-        public uint Tmms { get; set; }
+        public ulong Tmms { get; set; }
 
         [JsonProperty("tmst")]
         public uint Tmst { get; set; }
